# project/docs/apis/exergame-openapi.yaml
openapi: 3.0.3
info:
  title: Exergame API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /session:
    get:
      tags:
        - Session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "400":
          description: Bad Request
        "404":
          description: Not Found
    post:
      tags:
        - Session
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Session"
  /session/{id}:
    get:
      tags:
        - Session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
    put:
      tags:
        - Session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Session"
    delete:
      tags:
        - Session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
  /session/{childId}/last-month/total-hour:
    get:
      tags:
        - Session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
  /session/{childId}/recent:
    get:
      tags:
        - Session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: integer
            format: int64
...
components:
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        duration:
          type: string
...